check_version {regexplain}	R Documentation

Check if an updated version is available

Description

I included this because it can be difficult to tell if your RStudio Addins are up to date. I may add new features that you want but you won't hear about the updates. This function checks if an update is available, using GitHub tags. If an update is available, a modal dialog is shown when you start the regexplain gadget. This only happens once per R session, though, so feel free to ignore the message.

Usage

check_version(
  gh_user = "gadenbuie",
  gh_repo = "regexplain",
  this_version = packageVersion("regexplain")
)

Arguments

gh_user
GitHub user account

gh_repo
GitHub repo name

this_version
The currently installed version of the package

[Package regexplain version 0.2.2.9000 Index]