get_regex_library {regexplain}	R Documentation

Loads Regex Pattern Library

Description

Patterns sourced from Regex Hub are available at https://github.com/lukehaas/RegexHub and are copyright Luke Haas licensed under the MIT license available at https://github.com/lukehaas/RegexHub/commit/3ab87b5a4fd2817b42e2e45dcf040d4f0164ea37. Patterns source from qdapRegex are copyright Tyler Rinker and Jason Gray, licensed under the GPL-2 license.

Usage

get_regex_library()

[Package regexplain version 0.2.2.9000 Index]