regex {regexplain}	R Documentation

Extract matched groups from regexp

Description

Extract matched groups from regexp

Usage

regex(
  text,
  pattern,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE,
  global = TRUE
)

Arguments

text
Text to search

pattern
regexp

ignore.case
if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.

perl
logical. Should Perl-compatible regexps be used?

fixed
logical. If TRUE, pattern is a string to be matched as is. Overrides all conflicting arguments.

useBytes
logical. If TRUE the matching is done byte-by-byte rather than character-by-character. See ‘Details’.

global
If TRUE, enables global pattern matching

[Package regexplain version 0.2.2.9000 Index]