regexplain-package {regexplain}	R Documentation

regexplain: Rstudio Addin to Explain, Test and Build Regular Expressions

Description

A set of RStudio Addins to help interactively test and build regular expressions. Provides a Shiny gadget interface for interactively constructing the regular expression and viewing the results from common string-searching functions. The gadget interface includes a helpful regex syntax reference sheet and a library of common patterns.

Options

regexplain.no.check.version if TRUE or unset, the app will check for updates on the first run of the session.

regexplain.addin.max_lines controls the number of lines of text imported into the gadget, defaults to 100. Higher values will degrade performance.

Author(s)

Maintainer: Garrick Aden-Buie g.adenbuie@gmail.com

Other contributors:

Winston Chang (Author of textInput and textAreaInput fragments from shiny) [contributor]

Yihui Xie (Author of escape_html fragment from highr) [contributor]

RStudio (Copyright holder of included shiny fragments) [copyright holder]

See Also

Useful links:

https://github.com/gadenbuie/regexplain

Report bugs at https://github.com/gadenbuie/regexplain/issues

[Package regexplain version 0.2.2.9000 Index]