regexplain_addin {regexplain}	R Documentation

RegExplain Addin

Description

Addin for regexplain. If the selection does not contain multiple lines, the addin tries to evaluate the selection and coerce the result to a character string. Only unique lines are passed to the gadget to maximize screen space. The max number of lines returned is set by the option regexplain.addin.max_lines, with a default value of 100 lines. If evaluating the selection does not produce a character vector without errors, the original selection is returned.

Usage

regexplain_addin()

[Package regexplain version 0.2.2.9000 Index]