textInputCode {regexplain}	R Documentation

Modified Text Input

Description

Standard shiny::textInput() with additional width parameter, added code font style for the input text and with autocomplete, autocorrect, autocapitalize and spellcheck set to off or false.

Usage

textInputCode(
  inputId,
  label,
  value = "",
  width = NULL,
  placeholder = NULL,
  ...
)

Arguments

inputId
The input slot that will be used to access the value.

label
Display label for the control, or NULL for no label.

value
Initial value.

width
Width of shiny-input-container div.

placeholder
A character string giving the user a hint as to what can be entered into the control. Internet Explorer 8 and 9 do not support this option.

...
Extra elements to be included in the input-group div.

See Also

Other modified shiny inputs: textAreaInputAlt()

[Package regexplain version 0.2.2.9000 Index]