wrap_result {regexplain}	R Documentation

Wrap matches in HTML span tags to colorize via CSS

Description

Wrap matches in HTML span tags to colorize via CSS

Usage

wrap_result(x, escape = FALSE, exact = FALSE)

Arguments

x
Individual list item in list returned by regex()

escape
Escape HTML-related characters in text?

exact
Should the regex pattern be displayed as entered by the user into R console or source (default)? When TRUE, regex is displayed with the double \\\\ required for escaping backslashes in R. When FALSE, regex is displayed as interpreted by the regex engine (i.e. double \\\\ as a single \\).

[Package regexplain version 0.2.2.9000 Index]